+++ COLLADA2GLTF/extensions/o3dgc-compression/GLTF-Open3DGC.h
@@ -32,7 +32,7 @@ namespace GLTF
                             std::shared_ptr<JSONObject> floatAttributeIndexMapping,
                             GLTFAsset* asset);
     
-    void encodeDynamicVector(float *buffer, const std::string &path, size_t componentsCount, size_t count, GLTFAsset& asset);
+    void encodeDynamicVector(float *buffer, const std::string &path, long componentsCount, long count, GLTFAsset& asset);
     
     void setupAndWriteAnimationParameter(GLTFAnimation *cvtAnimation,
                                          const std::string& parameterSID,
